// src/types/Characters.ts
export type ServerBattleCharacter = {
  id: number;
  characterName: string;
  characterType: string;
  characterClass: string;
  characterSTR: number;
  characterDEX: number;
  characterCurrentHP: number;
  characterCurrentSHP: number;
  characterMaxHP: number;
  characterCurrentMP: number;
  characterCurrentSMP: number;
  characterMaxMP: number;
  characterMaxSMP: number;
  characterCurrentSP: number;
  characterMaxSP: number;
  characterStandingImageLink: string;
  characterIconImageLink: string;
};